+++ compiler-plugin-tests/src/test/java/io/ballerina/stdlib/email/compiler/CompilerPluginTest.java
@@ -158,6 +158,14 @@ public class CompilerPluginTest {
     }
 
     @Test
+    public void testReadonlyParameterPath() {
+        Package currentPackage = loadPackage("sample_package_12");
+        PackageCompilation compilation = currentPackage.getCompilation();
+        DiagnosticResult diagnosticResult = compilation.diagnosticResult();
+        Assert.assertEquals(diagnosticResult.errors().size(), 2);
+    }
+
+    @Test
     public void testListenerPreDeclaredListener() {
         Package currentPackage = loadPackage("sample_package_10");
         PackageCompilation compilation = currentPackage.getCompilation();